requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.6
django-redis==4.12.1
djangorestframework==3.12.2
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
psycopg2==2.8.6
pyparsing==2.4.7
pytz==2021.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.16.12
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
